Arrow head in the components diagram now scales properly
A value in pixels was added to width of the arrow head to make it
bigger. This changes the proportions of the arrow head when the
diagram is zoomed. Always do relative calculations base on the bounds
that are passed in.

Now a factor is used to increase the width.