Hyperlink style attribute (goto reference without cmd/ctrl)
There is a new style attribute "hyperlink" that can be used on reference cells. If it is set to true, you can click on the cell and follow the reference without holding the cmd/ctrl key. It also turns to blue underlined without cmd/ctrl.
Import "HyperlinkStyleAttribute" (model "com.mbeddr.mpsutil.hyperlink.editor") to make the style attribute available.