New dialog for choosing a node in the project
There is a class "TargetChooserDialog" that offers some static methods
for opening the dialog. You can choose between a blocking call, that
returns the selected node or a non-blocking call that accepts a callback.

There is a filter condition for the visible nodes, but at the moment all
ancestors of a node are shown. You cannot change the detail level of the
node hierarchy.