Performance: Vertical layout for the content of an implementation module
It was an indent layout before. The problem with indent layout is, that
in a subtree of indent layouts, the topmost collection does the
layouting for all descendants that also have an indent layout.

If indent layouts are used everywhere, the whole editor has to be
relayouted after every change.