Userstyles: Annotation on a style sheets allows the user to change the values
In Preferences->Editor->Colors & Fonts->mbeddr the user can change the
font color and style.

There is an example build script, that shows you how to create a plugin
that adds new predefined color schemes.

You can use the preferences dialog to create a new scheme and then use
an intention in the build script to turn it into a predefined one.