Workaround for a classloading issue in MPS
Project/Application plugins defined in runtime solutions are disposed to
late. You can register a listener for class reloading of modules. There
are the events "beforeClassesUnloaded" and "afterClassesLoaded".
The problem is, that plugins are loaded/unloaded asynchronously, so that
the dispose method of the plugin is called after the "afterClassesLoaded".
It should be called on the first event and never after the second.

The workaround is to listener for the first event and then doing the
dispose and not waiting for MPS to do the call.