Diagram: Autolayout: Edges can now connect to the nearest side of a box
There is a new option "auto layout: allow any side" for edge endpoints.
If it is set to true, you can move the box and the edge will connect to
the side with the shortest distance.

This works only if the edge is not connected to a port and only if you
layout edges only or move a box. It doesn't work when you relayout the
whole diagram.

Fixes #1252 (On a box that has no ports, the connectors should move around)