core: Some flags where not selectable and therefore and couldn't be deleted
The flags are in an editor component and at the usage of that component,
the style "selectable: false" was set. This was not a problem before
the inlining of the components, because the editor generator put a
collection around the flag cell (is required for making the "show if"
working).