tables: Performance optimization for tables without row headers
The algorithm for merging headers during the flattening transposes the
grid to apply the same algorithm on column and row headers. The
transpose is expensive and can be ommited if there are no row headers.

A better solution would be to implement a wrapper for the grid that
exchanges the x and y coordinates during the access. But this is more
effort, because it's easy to introduce new bugs.