mpsutil: http server takes port from property or env when in unit test
if the application is in unittest mode the http server now picks up the port
it uses from the env variable "MBEDDR_TEST_HTTP_PORT". If that variable is not
present or does not contain a valid port the property mbeddr.test.http.port is read
if it contains a valid TCP port, this port it used.

fixes #1420