Diagram: Model elements are not deleted during synchronisation anymore
The delete action now directly deletes the elements from the MPS model
and in the diagram they are removed on the next read sync.
This conforms more to the principle of a projectional editor.
In some cases elements where accidentally deleted from the model (most
of the time caused by an error in the editor definition). This change is
supposed to fix that, because elements can now only be removed by an
explicit user action.

See also #1470