TopDownLayout: More memory efficient size cache implementation
Instead of storing Size objects in a TreeMap, keys and values are stored
in a single int[].

One Size object requires 24 bytes and one TreeMap$Entry 40 bytes.
That’s 88 bytes for one entry. Now only 4*4=16 bytes are required.
Also the overhead for managing the objects is reduced. Only one array
object is required for the whole map.

See #1595