BuildAssistant: Provides context menu actions for building dependencies
In the context menu of models, modules and project you will find the two
actions “Build with Dependencies” and “Rebuild with Dependencies”.
They will collect all modules that the selection has a (transitive)
dependency on and start a build with them.

It is not required anymore to start a command line build. You can just
execute the action on the project/languages/solutions you want to work
with.

It will reduce the build time after pulling changes from a repository,
because you only have to rebuild the modules that you use.

Sometimes the build fails (because of class loading errors) and you have
to restart MPS, but you can just execute “Build with Dependencies” again
and MPS will skip all modules that are already built.