core: IVisibleElementsProvider migation preparation
Made visibleContentsOfType(concept<> targetConcept) virtual to allow
overwriting.

Deperecated the other methods in the API that take conceptNode paramters
and added overloads that use concep<> instead.

Changed the implementation of the conceptNode methods to call into the
new methods that use concept<>.

Reimplemented findByName with collection API rather than iterating the
sequence manually.

Added a comment explaning how a default implementation of visibleContentsOfType(conceptNode<>)
should look like given the fact that it is still abstract and needs to
be implemented but shouldn't duplicate the logic in visibleContentsOfType(concept<>)

Fixes #1846