mpsutil.targetchooser: target chooser always returns null
The problem was that the tree is disposed with the dialog meaning that after the dialog is closed
you could no longer get the selected object.

To go around this we save the selected object in the ok action so we can still get it after the
dialog is closed.