mpsutil.smodule:fixed GlobalModelAccess usage in addModelOperation
Because now GlobalModelAccess has no way to execute a command, we have
to provide a project repository to the add model operation.

For now I made the repository parameter optional so that we just don't
break the existing usages. When a repository is not provided we try to
get it from the opened projects.