doc: fix code removal in screenshots
Screenshots didn't respect the content removal specified in the
ModelContentAsScreenShotParagraph since we migrated to the logic
that takes screenshots from the intermediate model and not from
the original model.

This changes sets the editors hints for the intermediate model
correctly in the docs languages to allow the remove to work.

Only setting the hints wasn't enough because conditional editors
were disabled in the screenshooter code since the editor hints
for conditional editors were not set. The screenshooter now
always sets the conditional editor hint so that conditional