build: additional publishing to github packages
The build now publishes the master and maintenance/* branches to github
packages in addition to the normal itemis maven repository. The changes
are backwards compatible and don't require any changes on teamcity,
apart from providing the github logins data.

Read the inline comments carefully because some changes are none obvious
without. Mostly caused by the fact that we configure our subproject from
a parent project was tricky to get right.