build: add separate platform artifact
Added a new artifact to the mbeddr platform that doesn't repackage
mps-extensions. It exposes the dependency to mps-extensions via the pom.
This allows consumers of the artefact to decided which mps-extensions
version they would like to use.

This change breaks existing build scripts because of the new MPS build
solution that does only build the mbeddr platform. To fix this open the
build script in MPS and press F5 to fix all the references.

The artefact that repackaged mps-extensions is still present and remains
untouched. Down stream projects do not need to change their gradle scripts.