Generator performance (#2200)
As part of profiling efforts in a customer projects we identified a couple of hotspots during generation.

Most of them were typesystem related. To cache the expensive typesystem calls I introduced TypeCacheHelper which essentially applies memoization. It is used to wrap all call to the typesystem and will return a result from it's cache when it encounters the same node twice.

The new typesystem cache is used in:

handleArrayStructUnionInitExprAndAssignments
TransparentStructHelper
From my testing caching there doesn't break anything but please do some extensive testing in other codebases to see if this really holds.

The second hotspot we identified was sorting the module content. Here as well we apply caching. Since the sortied relies on function signatures the function to a given signature is cached in the Dependencies class. The class itself is also no longer static so that we can determine the lifecycle of the cache.

There are a couple of smaller optimisations in various places of the generator where multiple calls to .type are replaced with a local variable to store the result.

All in all these changes cut our generation by 40-50%.