Fix broken build
Due to some inexplicable Gradle behavior (probably a bug), when
`maven-publish` plugin is applied to `com.mbeddr` project, it causes the
subproject configuration in `subprojects` block to misbehave. Instead of
adding a repository to each subproject, the repository is added multiple
times to the parent.

Removing the `plugins` block from `com.mbeddr` fixes this.

Since `com.mbeddr:distribution` does not apply these plugins, publishing
repositories are now configured only when the `maven-publish` plugin is
applied.