mpsutil.actionsfilter: initialize the action manager before working with it
If we do not do this outside of the EDT thread, the class will complain because it can't be initialized on this thread.
Another similar issue in MPS: https://youtrack.jetbrains.com/issue/MPS-33812/Assertion-failed
This exception is very unlikely to happen anyway because MPS probably also needs to initialize this class somewhere.