fix: `com.mbeddr.core.base.ProjectScope` failed during a command line build
During test execution there was an additional temporary project open. That's why the strategy of
choosing the single open project didn't work.
The fallback logic of choosing the most recent focused window doesn't work in a headless environment
and even a normal IDE environment isn't a reliable strategy.

The new strategy is to find the project that contains the current module and as a fallback to
include all open projects.