fix: replace NavigationSupport with EditorNavigator
Migrate all usages of NavigationSupport to EditorNavigator (and one
usage to ProjectNavigator which is a similar helper class for locating
the node in the logical view).

NavigationSupport was often being used from a write lock, in 2025.1 this
leads to a deadlock. EditorNavigator presents a better API: it can be
called without holding a read/write lock and it accepts node references
instead of nodes.

Testing:

- hyperlink: open DummyConcept node in the
  com.mbeddr.mpsutil.hyperlink.sandbox.sandbox.manualTest model, click
  on 'A URL link to another node', MPS should not freeze.

Fixes #3175, fixes #3180.